requirements.txt
requests
functions-framework
PyJWT